Resolving issue "code quality - burn_away_slime function is better that copy pasting the same code"
In fact both instances of copied code should not have been there at all.

The first instance occured in the function that handles monsters falling into traps, with the result that observing a monster fall into a fire trap was enough to cure you  of sliming.

The second instance was redundant with an actual call to burn_away_slime a bit farther down in the function.
The code in the vicinity of the real call to burn_away_slime has been slightly revised for clairity.