Bug fix: Necronomicon
There is some wierd problem with the necronomicon when it is found on
a level rather than obtained via a wish, corpsenumber (aka spestudied,
same value, multiple names) is set to an apparently random number.
Fixing it in artifact.c the first time the book is read works around
the problem. Explicitly giving a default value in mkobj.c did not.