Merge of L's Monster Spells Patch
This merges in a modified form of L's Monster Spells Patch.

The primary modification is that I only slightly changed the vanilla spell casting lists, so that monsters will mostly still behave as they used to. As part of this, I retained both Finger of Death and Agrivate, which L had removed as useless. Turn to Stone actually seemed much nastier than Finger of Death to me, and I don't mind that FoD makes magic resistance almost requried. I kept Agrivation in the game because there are some levels in dNetHack where stealth is a good idea, and so a monster showing up to cast agrivation is a pretty tense thing. Most monsters that recieved signature spells as part of this patch still cast those signature spells, with the exception of the Gnomish Wizard.