Monster AI and Hod Sephiroth attack patch.
This is a merge between Grunt's Intelligent Pet Patch and dNetHack. It incorporates minimal fixes to allow the Intelligent Pet Patch to coexist with L's Monster Spells Patch. To wit: it converts all of the new spells to psi-bolt when cast by friendly creatures.

Mixed up in all of this is a small patch that causes the Hod Sepirah to use your equiped armor when calculating its AC.