Revisiting Wards
Implementing the magic stave wards at last.

These wards are carved into wooden weapons. This commit includes all the code needed to do that, as well as the code governing the effects of the staves.

Only 4 out of the 6 staves are currently implemented. The others don't do anything and shouldn't be learnable.

Also removes the mysterious reference to "engraving" on a weapon to name it, since we have actual rules for that now. I guess now it begs the question how you know the name if it isn't writen on the object though...