Bug fixes
So: Turns out, OR does not do the same thing as AND, and if you put an OR when you meant AND, your quest guardians will kill you.
-Translation: fixed a bug where all monsters (quest NPCs included) are generated hostile after killing Axus.

Also fixed a bug where the Tie-Dye Shirt was granting 2.5x carrying capacity instead of 1.5x. I THOUGHT it seemed a little too good....

Also adjusted the AC of the beholder up, as it isn't suposed to be that hard to kill.