General ward improvements, scrolls of ward and warding
Scroll of Ward
Always has an associated ward, its name appears in the inventory (ie, scroll of heptagram, scroll of pentagram).
When read, engraves its associated ward on the current square. If blessed, it is engraved at maximum reinforcement, if cursed, the ward is drawn in blood (and thus easly errased).

Scroll of Warding
When read, burns a chosen known ward into the current square. If blessed, it is burnt at maximum reinforcement. Cursed currently has no effect.

Wood weapons can be generated with a magical stave already on them.

Dwarves start knowing at least one magical stave.

Wizards and priests start knowing the correct number of wards.

Engraved rings and carved weapons can be wished for.
-Rings:
--Examples: Wish for "pentagram engraved ruby ring" or "pentagram ring of slow digestion"
--Wizard mode players always recive the wished-for engraving, if it is of a valid type.
--Other players recieve the wished-for engraving iff the ring would have been generated with a random useful engraving anyway.
--Pentagram engraved rings can be wished for, even though pentagram engraved rings aren't randomly generated.
-Weapons
--Examples: Wish for "toustefna carved staff"
--The specified stave will always be present. The player is spending a wish for a sub-standard weapon, after all!

NOTE:Just wishing for an "engraved ring" or a "carved staff" is not currently understood by the wishing code, you have to say what you want it engraved with.