Randomized monsters, and a few new tiles
Added shambling, stumbling, and wandering horrors: monsters with random statistics.

These are based on sporkhack's shambing horror, as implemented in unnethack.

Added code to save and restore random monsters, since apparently it is needed for dnethack, but not for sporkhack or unnethack?

New tiles: horror tiles, plus actual tile for uvuudaum and a new tile for the elder priest