Vampire Mod
Imported from UnNetHack and SLASH'EM.

Potions of blood were changed so that the blood is associated with a real creature.

If drunk, it gives a shot at all the resistances the an actual corpse of that creature would give.

Drinking potions of blood counts as eating meat, and may cause non-vampires to vomit.