Paradise Island/Sunken City Patch/Pirate improvements
This implements Fyr's Paradise Island YANI as an alternat of the Sunless Sea level. Both contain a magic lamp and some other treasure, and both feature water heavily.

Also adds the Sunken City as a third variant, a deep one city that also contains a magic lamp, assorted treasure, etc.

This patch also has a few improvements to the Pirate roll patch tangled up in it, namely that trees in the tropics (paradise island and the pirate quest) will drop different fruits and create parrots when kicked.

Trees in the zuggtmoy level will drop slime molds if kicked (even if the player has named their fruit to be something other than "slime mold")