Incantifier Patch: Tins
Incantifiers should be able to at least attempt to eat from tins.

If a tin is known to contain newt or aoa meat, the incantifier can eat from it.

If the contents of a tin are unknown, the incantifier can open it. The tin is discarded unless it turns out to contain newt or aoa meat.

A hallucinating incantifier will ALWAYS discard a tin on the grounds they can't eat its contents.