Iron ball of levitation
The quest leader gives the iron ball of levitation to you when you qualify for the quest

The ball gives luck, warning, stealth, level drain resistance, and increased carry cap when carried.

Invoke for levitation

It also protects you from punishment spells.

When used as a weapon, the ball does 2x damage