New scrolls! Antimagic, resistance, and consecration
Scrolls of Antimagic grant magic resistance but block your own spellcasting, for 400 turns.  4 scrolls are up in Vlad's throne room.

Scrolls of Resistance grant all 5 temporary energy resistances for 5 - 500 turns.  Typically, 3 such scrolls may be found in the quest leader's treasure chest.

The Scroll of Consecration creates an altar.  Be sure to read it on a regular floor square, or your god will get angry at you and you won't get the altar.  The scroll of consecration can be found in the black temple.  You can't write or wish for scrolls of this type.