Convict quest changes
The plot of the Convict Patch is that the warden has somehow become possessed by demons, and is leading some sort of summoning ritual in the basement of Castle Waterdeep.  This expands on that plot.
-Fill the basement with Inmate corpses
-Create a second set of quest monsters for the basement
--Daughters of Bedlam, quasits, and random demons.
-The warden has Kyton equipment
--Note: I've been meaning to add Kytons, and this would be a good place to feature them.

Also add a minor artifact system that allows the addition of named magic items with special properties in .des files
-Use it to add some Waterdeep-themed gear
-Remove Magic Resistance from the Spoon, in favor of the Waterdeep gear.
--The spoon grants slotless Free Action instead

Also adds a water-damage magic weapon property, which the warden's old sword sometimes gets.