Bones files and trephination kit groundwork.
Glyphs can be stored on bone shards.

Reading a shard teaches the glyph (if you are below limit and don't have it already)

In bones files, glyphs are converted into bone shards and dropped or given to monster you rose as.

Also, in bones files binder seals scatter around the level (FINALLY!!1!)